audtool: Accept a floating-point argument to seek functions. Closes: #95.

+++ src/audtool/handlers_playback.c
@@ -107,7 +107,7 @@ void playback_seek(gint argc, gchar **argv)
 		exit(1);
 	}
 
-	audacious_remote_jump_to_time(dbus_proxy, atoi(argv[1]) * 1000);
+	audacious_remote_jump_to_time(dbus_proxy, atof(argv[1]) * 1000);
 }
 
 void playback_seek_relative(gint argc, gchar **argv)
@@ -121,7 +121,7 @@ void playback_seek_relative(gint argc, gchar **argv)
 	}
 
 	oldtime = audacious_remote_get_output_time(dbus_proxy);
-	diff = atoi(argv[1]) * 1000;
+	diff = atof(argv[1]) * 1000;
 	newtime = oldtime + diff;
 
 	audacious_remote_jump_to_time(dbus_proxy, newtime);
